backend/requirements.txt
# app
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
email-validator==2.0.0.post2
# db
databases[postgresql]==0.8.0
SQLAlchemy==1.4.49
alembic==1.11.3
psycopg2-binary==2.9.7
# dev
pytest==7.4.1
pytest-asyncio==0.21.1 # provides utilities for testing asynchronous code
httpx==0.24.1 # provides an async request client for testing endpoints
asgi-lifespan==2.1.0 # allows testing async applications without having to spin up an ASGI server